split 3den config
