+++ TER_VASS/addons/3den/Attributes/Base.hpp
@@ -1,0 +1,28 @@
+class Default;
+class Title: Default
+{
+	class Controls
+	{
+		class Title: ctrlStatic
+		{
+			style = 1;
+			x = 0;
+			w = ATTRIBUTE_TITLE_W * GRID_W;
+			h = ATTRIBUTE_CONTENT_H * GRID_H;
+			colorBackground[] = {0,0,0,0};
+		};
+	};
+};
+class TitleWide: Default
+{
+	class Controls
+	{
+		class Title: ctrlStatic
+		{
+			x = ATTRIBUTE_CONTENT_H * GRID_W;
+			w = (ATTRIBUTE_TITLE_W + ATTRIBUTE_CONTENT_W - ATTRIBUTE_CONTENT_H) * GRID_W;
+			h = ATTRIBUTE_CONTENT_H * GRID_H;
+			colorBackground[] = {0,0,0,0};
+		};
+	};
+};
